extracted SourceRow into separate component;
